--- src/types/config.ts
@@ -1,7 +1,0 @@
-export type ConfigSection = {
-  title: string
-  btnText: string
-  btnBg: string
-  btnHoverBg: string
-  onClick: () => void
-}
